YL_YTDownloader-main/requirements.txt
customtkinter==5.2.1
pymongo==4.6.1
bcrypt==4.1.2
yt-dlp==2023.12.30
Pillow==10.2.0
requests==2.31.0
python-jose==3.3.0
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
google-auth==2.27.0
pyotp==2.9.0
qrcode==7.4.2
matplotlib==3.8.2
pandas==2.2.0
reportlab==4.1.0
python-magic==0.4.27
ffmpeg-python==0.2.0
pytest==8.0.0
black==24.1.1
flake8==7.0.0